refactor(product): add sub directories for product selectors
