+++ libs/product/src/selectors/all-selectors.selectors.ts
@@ -1,8 +1,8 @@
 import { DaffProduct } from '../models/product';
-import { DaffProductPageMemoizedSelectors, getDaffProductPageSelectors } from './product.selectors';
-import { DaffBestSellersMemoizedSelectors, getDaffBestSellersSelectors } from './best-sellers.selectors';
-import { DaffProductEntitiesMemoizedSelectors, getDaffProductEntitiesSelectors } from './product-entities.selectors';
-import { DaffProductGridMemoizedSelectors, getDaffProductGridSelectors } from './product-grid.selectors';
+import { DaffProductPageMemoizedSelectors, getDaffProductPageSelectors } from './product/product.selectors';
+import { DaffBestSellersMemoizedSelectors, getDaffBestSellersSelectors } from './best-sellers/best-sellers.selectors';
+import { DaffProductEntitiesMemoizedSelectors, getDaffProductEntitiesSelectors } from './product-entities/product-entities.selectors';
+import { DaffProductGridMemoizedSelectors, getDaffProductGridSelectors } from './product-grid/product-grid.selectors';
 import { DaffProductFeatureMemoizedSelector, getDaffProductFeatureSelector } from './product-feature.selector';
 
 export interface DaffProductAllSelectors<T extends DaffProduct = DaffProduct> extends
