+++ libs/product/src/selectors/product-grid/product-grid.selectors.ts
@@ -1,9 +1,9 @@
 import { createSelector, MemoizedSelector } from '@ngrx/store';
 
-import { DaffProductGridReducerState } from '../reducers/product-grid/product-grid-reducer-state.interface';
-import { DaffProductReducersState } from '../reducers/product-reducers-state.interface';
-import { getDaffProductFeatureSelector } from './product-feature.selector';
-import { DaffProduct } from '../models/product';
+import { DaffProductGridReducerState } from '../../reducers/product-grid/product-grid-reducer-state.interface';
+import { DaffProductReducersState } from '../../reducers/product-reducers-state.interface';
+import { getDaffProductFeatureSelector } from '../product-feature.selector';
+import { DaffProduct } from '../../models/product';
 
 export interface DaffProductGridMemoizedSelectors<T extends DaffProduct = DaffProduct> {
 	selectProductGridState: MemoizedSelector<object, DaffProductGridReducerState<T>>;
