+++ libs/product/src/selectors/product/product.selectors.ts
@@ -1,9 +1,9 @@
 import { createSelector, MemoizedSelector } from '@ngrx/store';
 
-import { DaffProductReducersState } from '../reducers/product-reducers-state.interface';
-import { getDaffProductFeatureSelector } from './product-feature.selector';
-import { DaffProductReducerState } from '../reducers/product/product-reducer-state.interface';
-import { DaffProduct } from '../models/product';
+import { DaffProductReducersState } from '../../reducers/product-reducers-state.interface';
+import { getDaffProductFeatureSelector } from '../product-feature.selector';
+import { DaffProductReducerState } from '../../reducers/product/product-reducer-state.interface';
+import { DaffProduct } from '../../models/product';
 
 export interface DaffProductPageMemoizedSelectors<T extends DaffProduct = DaffProduct> {
 	selectSelectedProductState: MemoizedSelector<object, DaffProductReducerState>;
